Allow DoubleEntryPoint bot reset on same instance
I saw that in the original PR to create this level there was some discussion around leaving instance usable in case of validation failure. Raising this PR assuming that instance reuse-ability is a desired property for this level but once we set a detection bot we can not set another one with different logic. If we can not reset the detection bot then the submission will keep using old failed bot and will keep failing.